tools/calculate_date_difference/requirements.txt
pytz==2024.1
python-dateutil==2.9.0 